Consistently intern the action function names into the METAPEG package, not the cl-user package.  I've not tested if this broke the bootstrapping.
That change is part of a move toward an alternative model of what package the
actions are authored in.  The goal is to have them authored in the package that
is current when the code for their parser is generated.  That's working now,
for me, but I've yet to commit an example here.

Change the logging done when an action function appears to be
undefined so it shows actual error, thus when the undefined function
is a function used in the action you can see that in the log.

Port this to CCL (aka openmcl).  The parser catches errors, and those
errors tend to be implementation specific.  The original had SBCL
cases.  This adds the analagous cases for CCL.  Sadly those aren't as
exacting as the SBCL errors.

The actions names are spelled slightly differently for estetics
i.e. |metapeg_action999| is now METAPEG-ACTION999

Finally there are some misc changes to resolve compiler warnings.
Some declare ignore, or removing unused bindings of the error in the
handler-cases.